refactor(analysis): make analysis_sql a package with a shared core
First step of #1089. analysis_sql.py becomes analysis_sql/__init__.py and the
machinery no single family owns moves to analysis_sql/shared.py: the OFFSET 0
pull-up fence and the two aliases built on it, the measured source and mask
ceilings with the benchmark tables that justify them, the antimeridian helper
and its threshold set, and the mask parser.

Every one of those would become a drift surface if a family kept its own copy,
which is the same reason the module lives in platform/ rather than in either
caller. Three placements the issue did not name, decided here:

- render_mask_expr is central, not in overlay with its only two callers. It is
  the injection boundary the package docstring describes — the one place
  untrusted GeoJSON becomes SQL text — so it should be findable by asking that
  question, not by knowing which operations happen to take a drawn mask.
- INTERNAL_ALIAS_PREFIX is central. The whole design of a reserved prefix over
  a list of names is that it covers an alias added later; owned by overlay, an
  alias from another family reads as out of scope, which is the list-falls-
  behind failure the prefix replaced.
- NON_GROUPABLE_COLUMN_TYPES is central. It is a PostgreSQL capability fact
  with no rendered statement behind it, and its consumer set already crossed a
  family boundary once inside one release (#1099 took intersect off it).

The façade declares __all__ rather than relying on the imports it happens to
bind: an F401 autofix would otherwise read every re-export as dead, and it is
also the list a reviewer diffs against the pre-split module.

No rendered statement changed — the byte-identity harness passes all 94 cases
and the façade exports every symbol the single file defined. Ratchet retargeted
to the façade and re-measured: 1255 -> 1024, exact.
